Added New scripts
Player objects includes a rigidbody to it and also a box collider which
will not be in use for now.

Player Controller Script that will control the player throughout

Player movement in player controller script allows the player to move
either with the W,A,S,D or else with the arrow keys.

Camera Follow Script that will follow the player according to the
keys/direction set by the user